// src/enums/valid-extensions.enum.ts
export enum ValidExtensions {
  // Images
  JPG = 'jpg',
  JPEG = 'jpeg',
  PNG = 'png',
  GIF = 'gif',

  // Documents
  PDF = 'pdf',
  DOCX = 'docx',
  XLSX = 'xlsx',
  PPTX = 'pptx',
  ODT = 'odt',
  ODS = 'ods',
  ODP = 'odp',
  TXT = 'txt',
  RTF = 'rtf',

  // Video
  AVI = 'avi',
  MP4 = 'mp4',
  MOV = 'mov',
  FLV = 'flv',
  MKV = 'mkv',

  // Compression
  ZIP = 'zip',
  RAR = 'rar',
  TAR = 'tar',
  '7Z' = '7z',
}